attemp Coverage Increase, clean up NavigationActions
