+++ app/src/main/java/ch/epfl/cs311/wanderwave/ui/navigation/NavigationActions.kt
@@ -13,15 +13,14 @@ object Route {
   const val LOGIN = "login"
   const val MAIN = "main"
   const val TRACK_LIST = "trackList"
-  const val THEME = "theme"
 }
 
 // Top level destination
 data class TopLevelDestination(val route: String, val icon: ImageVector, val textId: Int)
 
-class NavigationActions(val navController: NavHostController) {
+class NavigationActions(navController: NavHostController) {
 
-  val navigationController = navController
+  private val navigationController = navController
 
   fun navigateTo(destination: TopLevelDestination) {
     navigationController.navigate(destination.route) {
@@ -36,20 +35,14 @@ class NavigationActions(val navController: NavHostController) {
       restoreState = true
     }
   }
-
+  /*
   fun goBack() {
     navigationController.popBackStack()
   }
-
+  */
   val navigateToLogin: () -> Unit = {
     navigationController.navigate("login") { popUpTo(Route.LOGIN) { inclusive = true } }
   }
-
-  val navigateToCreateToDo: () -> Unit = { navigationController.navigate("createToDo") }
-  val navigateToEditToDo: (String) -> Unit = { taskId ->
-    navigationController.navigate("editToDo/$taskId")
-  }
-  val navigateToOverview: () -> Unit = { navigationController.navigate("overview") }
 }
 
 val TOP_LEVEL_DESTINATIONS =
@@ -58,5 +51,4 @@ val TOP_LEVEL_DESTINATIONS =
         TopLevelDestination(route = Route.LOGIN, icon = Icons.Filled.AccountBox, textId = 2),
         TopLevelDestination(route = Route.TRACK_LIST, icon = Icons.Filled.List, textId = 3),
         TopLevelDestination(route = Route.MAIN, icon = Icons.Filled.List, textId = 4),
-        TopLevelDestination(route = Route.THEME, icon = Icons.Filled.List, textId = 5),
     )
