+++ app/src/main/java/ch/epfl/cs311/wanderwave/ui/screens/MainPlaceHolder.kt
@@ -19,3 +19,63 @@ fun MainPlaceHolder(navigationActions: NavigationActions) {
         }
   }
 }
+
+/*
+@Composable
+fun ThemeScreen() {
+  Column(
+      horizontalAlignment = Alignment.CenterHorizontally,
+      verticalArrangement = Arrangement.Center,
+      modifier = Modifier.fillMaxSize().testTag("themeScreen")) {
+        Text(
+            text = "Launch Screen",
+            modifier = Modifier.padding(16.dp),
+            fontWeight = FontWeight(700),
+            fontSize = 40.sp,
+            color = MaterialTheme.colorScheme.onBackground)
+        Box(
+            modifier =
+                Modifier.fillMaxWidth()
+                    .padding(16.dp)
+                    .border(2.dp, MaterialTheme.colorScheme.primary, RoundedCornerShape(16.dp))
+                    .padding(16.dp)) {
+              Text(
+                  text = "Primary",
+                  modifier = Modifier.align(Alignment.Center),
+                  fontWeight = FontWeight(700),
+                  fontSize = 40.sp,
+                  color = MaterialTheme.colorScheme.primary)
+            }
+
+        Box(
+            modifier =
+                Modifier.fillMaxWidth()
+                    .padding(16.dp)
+                    .clip(RoundedCornerShape(16.dp))
+                    .background(MaterialTheme.colorScheme.surface)
+                    .padding(16.dp)) {
+              Text(
+                  text = "primary on surface",
+                  modifier = Modifier.align(Alignment.Center),
+                  fontWeight = FontWeight(700),
+                  fontSize = 30.sp,
+                  color = MaterialTheme.colorScheme.primary)
+            }
+
+        Box(
+            modifier =
+                Modifier.fillMaxWidth()
+                    .padding(16.dp)
+                    .clip(RoundedCornerShape(16.dp))
+                    .background(MaterialTheme.colorScheme.primaryContainer)
+                    .padding(16.dp)) {
+              Text(
+                  text = "Primary Container",
+                  modifier = Modifier.align(Alignment.Center),
+                  fontWeight = FontWeight(700),
+                  fontSize = 40.sp,
+                  color = MaterialTheme.colorScheme.onPrimaryContainer)
+            }
+      }
+}
+ */
